requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
google-generativeai